+++ BlogposterCMS/mother/modules/plainSpace/config/adminPages.js
@@ -81,7 +81,7 @@ module.exports.ADMIN_PAGES = [
         sidebar: 'default-sidebar',
         inheritsLayout: true
       },
-      widgets: []
+      widgets: ['widgetList']
     }
   },
   {